Simplify default targets
We now only provide Stdout, and don't provide a convenience function for
logging to an arbitrary os.File. In practice, I don't think this will be
missed, since the logic for doing things like log rotation is a teensy
bit tricky and didn't fit into the API I wanted to provide here.